+++ veecle-os-examples/common/src/actors/mod.rs
@@ -1,7 +1,6 @@
 //! Actors shared between multiple examples.
 
-// `telemetry` includes `alloc`, but requiring both is more explicit and less likely to get mistakenly refactored.
-#[cfg(all(feature = "alloc", feature = "telemetry"))]
+#[cfg(feature = "alloc")]
 pub mod alloc;
 
 pub mod ping_pong;