+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Veecle OS
 
 * **breaking** Telemetry functionality is now always compiled into the runtime.
-  The `veecle-telemetry` feature flag has been removed from `veecle-os-runtime`.
+  The `veecle-telemetry` feature flag has been removed from `veecle-os-runtime` and `telemetry` feature flag has been removed from `veecle-os`.
   Use the `telemetry-enable` feature flag on `veecle-os` to control telemetry behavior.
 * Added `ThreadAbstraction` trait to OSAL for querying current thread id.
 * Fixed `veecle_os::telemetry::instrument` macro to automatically resolve correct crate paths for the facade.